requirement.txt
python-telegram-bot
openrouter
langchain 
langgraph 
requests 
apscheduler 
praw
langchain-openai
langchain-openrouter
langchain-community 
duckduckgo-search